Revert previous rename
