+++ GUI/src/components/ApiEndpointCard/Endpoints/RequestVariables/index.tsx
