+++ GUI/src/components/ApiEndpointCard/Endpoints/ResqlComponent/index.tsx
