+++ GUI/src/components/index.tsx
@@ -34,11 +34,11 @@ import Track from "./Track";
 import ExclamationBadge from "./ExclamationBadge";
 import OutputElementBox from "./OutputElementBox";
 import FlowElementsPopup from './FlowElementsPopup'
-import ApiEndpointCard from "./ApiEndpointCards";
-import EndpointOpenAPI from "./ApiEndpointCards/Endpoints/OpenAPI";
-import EndpointCustom from "./ApiEndpointCards/Endpoints/Custom";
-import EndpointResqlComponent from "./ApiEndpointCards/Endpoints/ResqlComponent";
-import RequestVariables from "./ApiEndpointCards/Endpoints/RequestVariables";
+import ApiEndpointCard from "./ApiEndpointCard";
+import EndpointOpenAPI from "./ApiEndpointCard/Endpoints/OpenAPI";
+import EndpointCustom from "./ApiEndpointCard/Endpoints/Custom";
+import EndpointResqlComponent from "./ApiEndpointCard/Endpoints/ResqlComponent";
+import RequestVariables from "./ApiEndpointCard/Endpoints/RequestVariables";
 
 export {
   Collapsible,
